move ack conversion function to own file
